Move cluster caching utils to separate package
